fastapi/auth/oauth2.py
from fastapi import HTTPException, status
from fastapi.param_functions import Depends
from fastapi.security import OAuth2PasswordBearer
from typing import Optional
from datetime import datetime, timedelta
from jose import jwt
from jose.exceptions import JWTError
from db.databases import get_db
from sqlalchemy.orm.session import Session
from db.db_user import db_get_user, db_get_username


# define token retrieval endpoint
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

# secret key used to sign the token
# new secret $openssl rand -hex 32
SECRET_KEY = 'bb92b11a4fdaf2b38169c0ac70850d39b8c3763fa053af25c80e71379c7768f2'
ALGORITHM = 'HS256'
ACCESS_TOKEN_EXPIRE_MINUTES = 30


# JSON web token
# https://jwt.io/
def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail='Could not validate credentials',
        headers={'WWW-Authenticate': 'Bearer',
                 }
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get('sub')
        if username is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception

    user = db_get_username(db, username)

    if user is None:
        raise credentials_exception

    return user
