fastapi/requirements.txt
ipython
requests
pymongo
motor
pandas
sqlalchemy
psycopg2
autopep8
ipykernel
fastapi
redis
uvicorn
strawberry-graphql
websockets
pydantic
passlib
bcrypt
python-multipart
python-jose
aiofiles
pytest
jinja2